% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/agreguj_wskazniki_ewd.R
\name{agreguj_wskazniki_ewd}
\alias{agreguj_wskazniki_ewd}
\title{Agregowanie wskaźników EWD}
\usage{
agreguj_wskazniki_ewd(dane, poziom = NULL, grupujPoLatach = TRUE,
  funkcjaGrupujaca = NULL, tylkoWyswietlane = NA, czyPomin = TRUE)
}
\arguments{
\item{dane}{ramka danych zwracana przez funkcję \link{pobierz_wartosci_wskaznikow_ewd}.}

\item{poziom}{ciąg znaków definiujący poziom agregacji: "gmina", "powiat" lub "województwo" lub NULL,
jeśli podany został parametr 'funkcjaGrupujaca'}

\item{grupujPoLatach}{wartość logiczna (TRUE lub FALSE) wskazująca, czy przy wyróżnianiu grup oprócz TERYTu
ma być uwzględniona również zmienna opisująca okres, dla którego wyliczony został wskaźnik}

\item{funkcjaGrupujaca}{fukcja przyjmująca dwa parametry: teryt oraz rok_do egzaminu, a zwracająca jednokolumnową całkowitoliczbową ramkę danych,
która określa grupowanie wskaźników ewd. W zwracanej ramce danych przez agreguj_wskazniki_ewd()
jedna kolumna będzie zawierać wynik funkcji grupującej (jeżeli grupujPoLatach=FALSE) lub sześć ostatnich cyfr wyniku tej funkcji (jeżeli grupujPoLatach=TRUE).}

\item{tylkoWyswietlane}{parametr nie oprogramowany}

\item{czyPomin}{parametr nie oprogramowany}
}
\value{
data frame
}
\description{
Funkcja agreguje wskaźniki EWD.
}
\examples{
\dontrun{
dane = pobierz_wartosci_wskaznikow_ewd('T', c(2013, 2012), dodatkoweInfo = FALSE, jst = "^2605|^2601", opisoweNazwy = FALSE, tylkoWyswietlane = FALSE, tylkoWskDoPrezentacji = FALSE)
agreguj_wskazniki_ewd(dane, "powiat", grupujPoLatach = FALSE)
agreguj_wskazniki_ewd(dane, "powiat", grupujPoLatach = TRUE)
}
}

