% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/agreguj_wskazniki_ewd.R
\name{konwertuj_nazwy_na_opisowe}
\alias{konwertuj_nazwy_na_opisowe}
\title{Kowersja nazw na opisowe}
\usage{
konwertuj_nazwy_na_opisowe(nazwyKolumn, revert = FALSE)
}
\arguments{
\item{nazwyKolumn}{nazwy kolumn ramki danych}

\item{revert}{parametr logiczny. Jeżeli revert wynosi TRUE to nazwy opisowe są zmienane na zwykłe.
Domyślna wartość FALSE powoduje zmianę nazw zwykłych na nazwy opisowe.}
}
\value{
wektor ciągów znakowych
}
\description{
Funkcja konwertuje nazwy kolumn ramki danych zwracanych przez funkcję \code{\link{pobierz_wartosci_wskaznikow_ewd}}.
}

