% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobierz_baze_szkol.R
\name{pobierz_baze_szkol}
\alias{pobierz_baze_szkol}
\title{Aktualizacja bazy danych szkol.}
\usage{
pobierz_baze_szkol(typySzkol, zapis = NULL)
}
\arguments{
\item{typySzkol}{opcjonalny wektor tekstowy z typami szkół, które mają zostać
zwrócone (lub NULL - zwraca informacje o wszystkich szkołach)}

\item{zapis}{opcjonalnie nazwa pliku, do którego zostaną zapisane pobrane dane
(w formacie csv)}
}
\value{
data frame (niewidocznie)
}
\description{
Funkcja pobiera dane o szkołach, potrzebne do dokonania aktualizacji bazy
szkół (wykonuje zrzut bazy szkół).
}

