% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pobierz_dane_wyskalowane.R
\name{pobierz_dane}
\alias{pobierz_dane}
\title{Pobieranie danych maturalnych}
\usage{
pobierz_dane(typSzkoly, lata, wydl = 0, src = NULL, demo = FALSE)
}
\arguments{
\item{typSzkoly}{ciąg znaków: 'gimn.', 'LO' lub 'T'}

\item{lata}{wektor liczb całkowitych - lata przeprowadzenia egzaminu
"na wyjściu", które mają znaleźć się w pobranych danych}

\item{wydl}{liczba całkowita - o tyle maksymalnie lat wstecz względem
standardowej długości toku kształcenia będzie się cofać się przy przyłączaniu
wyników egzaminu "na wyjściu"}

\item{src}{opcjonalnie uchwyt źródła danych dplyr-a do bazy EWD (jeśli chcemy
połączyć się z bazą z uprawnieniami wyższymi niż domyślne)}

\item{demo}{domyślnie FALSE. Jeżeli TRUE to pobiera tylko trzy konkretne
rekordy}
}
\value{
data frame
}
\description{
Funkcja pobiera z bazy dane maturalne dla wybranych roczników.
}
\examples{
\dontrun{
  src = polacz()
  ret = pobierz_dane_maturalne(src, 2012:2014, TRUE)
}
}

