% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_dane_do_skracania}
\alias{pobierz_dane_do_skracania}
\title{Pobieranie danych przefiltrowanych pod kątem wyliczania skrótów skal ocen}
\usage{
pobierz_dane_do_skracania(rodzajEgzaminu, czescEgzaminu, rokEgzaminu,
  idSkali = NULL, zrodloDanychODBC = "EWD")
}
\arguments{
\item{rodzajEgzaminu}{rodzaj egzaminu}

\item{czescEgzaminu}{część egzaminu}

\item{rokEgzaminu}{rok egzaminu}

\item{idSkali}{numer skali}

\item{zrodloDanychODBC}{opcjonalnie nazwa źródła danych ODBC, dającego dostęp do bazy (domyślnie "EWD")}
}
\value{
data frame
}
\description{
Funkcja pobiera dane z określonej części egzaminu. Następnie z tych danych usuwani są uczniowie, którzy przystępowali do egzaminu wielokrotnie lub uczęszczali do szkoły dla dorosłych, specjalnych lub przyszpitalnych. Usunięto także uczniów, których odział był oznaczony cyfrą.
}

