% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pobierz_dane_wyskalowane.R
\name{pobierz_dane_gimnazjalne}
\alias{pobierz_dane_gimnazjalne}
\title{Pobieranie danych gimnazjalnych}
\usage{
pobierz_dane_gimnazjalne(src, rok_gim, demo = FALSE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a do bazy EWD.
szkoły zwrócone zostaną tylko najświeższe dane w ramach tego okresu)}

\item{rok_gim}{wektor lat egzaminu}

\item{demo}{domyślnie FALSE. Jeżeli TRUE to pobiera tylko trzy konkretne rekordy (zniknie z docelowej wersji).}
}
\value{
data frame
}
\description{
Funkcja pobiera z bazy wyskalowane dane gimnazjalne dla wybranych roczników.
}
\examples{
\dontrun{
src = polacz(user = "user", password = "pass")
rok_matury = 2012:2014
wydl=1
rok_gim = (min(rok_matury)-3-wydl):(max(rok_matury)-3)
ret = pobierz_dane_gimnazjalne(src, rok_gim, TRUE)
}
}

