% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_dane_kontekstowe.R
\name{pobierz_dane_kontekstowe}
\alias{pobierz_dane_kontekstowe}
\title{Pobieranie danych kontekstowych o uczniach}
\usage{
pobierz_dane_kontekstowe(src, rodzajEgzaminu)
}
\arguments{
\item{src}{połączenie z bazą danych IBE zwracane przez funkcję
\code{\link[ZPD]{polacz}}}

\item{rodzajEgzaminu}{ciąg znaków - rodzaj egzaminu}
}
\value{
data.frame
}
\description{
Pobiera z bazy danych zmaterializowany zbiór danych zawierający wszelkie
infromacje kontekstowe potrzebne przy obliczaniu EWD.

Jeden wiersz zwracanej ramki danych opisuje pierwsze/ostatnie (patrz niżej)
przystąpienie ucznia do wskazanego rodzaju egzaminu, a informacje dotyczące
poszczególnych części egzaminu (na chwilę obecną tylko informacja o byciu
laureatem z danego przedmiotu) znajdują się w kolumnach prefiksowanych
zgodnie z tablicą sl_czesci_egzaminow w bazie.

Jeśli informacje o szkole, klasie, numerze w dzienniku lub dysleksji nie są
dla danego ucznia zgodne w różnych częściach egzaminu, wtedy dana informacja
(szkoła, klasa, dysleksja) zastępowana jest brakiem danych.

W zależności od poziomu uprawnień połączenia z bazą danych funkcja
automatycznie pobiera lub nie informacje o kodzie klasy i numerze ucznia w
dzienniku oraz o jego wieku, który podawany jest jako liczba ukończonych
miesięcy życia w dniu pisania najwcześniej przeprowadzonej części egzaminu
(w danym roku).

Dane pobierane są w postaci zmaterializowanej z powodów wydajnościowych -
zapytanie realizujące całą niezbędną funkcjonalnośc po stronie bazy danych
wykonywałoby się wieki.
}
\details{
Oprócz innych informacji funkcja zwraca również zmienne logiczne opisujące,
czy dane osoby należą do "wzorcowych" populacji, wykorzystywanych do
estymacji parametrów modeli skalowania i/lub modeli EWD. Są to zmienne
\code{populacja_we} i \code{populacja_wy}. Członkowie populacji "wzorcowych"
definiowani są jako zdający:
\itemize{
  \item{Którzy uczyli się w szkołach dla młodzieży, nie będących szkołami
        specjalnymi, przyszpitalnymi, itp.}
  \item{Którzy podchodzili do egzaminu:
        \itemize{
          \item{Po raz pierwszy - przy skalowaniu i odnośnie wyników
                "na wyjściu" w modelach EWD - zmienna \code{populacja_wy}.}
          \item{Po raz ostatni (odnotowany) - odnośnie wyników "na wejściu"
                w modelach EWD - zmienna \code{populacja_we}.}
        }
       }
  \item{Dodatkowo, tylko dla zmiennej \code{populacja_wy}, a więc
        w odniesieniu do skalowania i wyników "na wyjściu" w modelach EWD,
        populacja "wzorcowa" zawężana jest do zdających:
        \itemize{
          \item{Którzy należą do "głównej" kohorty wiekowej dla danego
                egzaminu, lub kohort o rok lub dwa lata starszych lub
                młodszych, tzn.:
                \itemize{
                  \item{Dla ucznów SP w wieku 124-183 miesięcy.}
                  \item{Dla uczniów gimn. w wieku 160-219 miesięcy.}
                  \item{Dla uczniów LO w wieku 196-255 miesięcy.}
                  \item{Dla uczniów T w wieku 208-267 miesięcy.}
                }
                Przy tym informacja o wieku nie jest uwzględniana w przypadku
                sprawdzianu z lat 2002-2004, egzaminu gimnazjalnego z lat
                2002-2007 i matury z lat 2002-2008, bo informacje o nim są
                w tych latach niedostępne.
          }
          \item{W przypadku matury brane są pod uwagę wyłącznie osoby,
                których kod klasy nie jest liczbą. Wynika to z faktu, że
                liczbowe oznaczenia klas często wykorzystywane są przez OKE
                dla oznaczenia osób, które nie były uczniami danej szkoły,
                ale pisały w niej maturę.}
        }
       }
}
Dodatkowe uwagi:
\itemize{
  \item{Informacje o wieku i klasie są użyteczne tylko wtedy, jeśli 
        przekazane do funkcji połączenie z bazą posiada poziom uprawnień,
        który umożliwia ich pobranie.}
  \item{Przed przystąpieniem do skalowania na populacji "wzorcowej" konieczne
        jest jeszcze usunięcie z danych z wynikami surowymi egzaminów
        punktacji za zadania przypisanej laureatom. Musi to jednak zostać
        zrobinone oddzielnie dla każdej części egzaminu (na podstawie
        zwróconych przez tę funkcję informacji o byciu laureatem).}
}
}
