% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pobierz_dane_wyskalowane.R
\name{pobierz_dane_maturalne}
\alias{pobierz_dane_maturalne}
\title{Pobieranie danych maturalnych}
\usage{
pobierz_dane_maturalne(src, rok_mat, demo = FALSE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a do bazy EWD.
szkoły zwrócone zostaną tylko najświeższe dane w ramach tego okresu)}

\item{rok_mat}{wektor lat egzaminu}

\item{demo}{domyślnie FALSE. Jeżeli TRUE to pobiera tylko trzy konkretne rekordy (zniknie z docelowej wersji).}
}
\value{
data frame
}
\description{
Funkcja pobiera z bazy dane maturalne dla wybranych roczników.
}
\examples{
\dontrun{
src = polacz(user = "user", password = "pass")
ret = pobierz_dane_maturalne(src, 2012:2014, TRUE)
}
}

