% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobierz_parametry_skalowania.R
\name{pobierz_parametry_skalowania}
\alias{pobierz_parametry_skalowania}
\title{Pobieranie parametrow skalowan}
\usage{
pobierz_parametry_skalowania(skala, skalowanie = NULL, doPrezentacji = NA,
  parametryzacja = "baza")
}
\arguments{
\item{skala}{id_skali (liczba naturalna) lub ciąg znaków z wyrażeniem
regularnym, do którego ma pasować opis skali}

\item{skalowanie}{opcjonalnie numer skalowania (liczba naturalna) lub ciąg
znaków z wyrażeniem regularnym, do którego ma pasować opis skalowania}

\item{doPrezentacji}{wartość logiczna - czy zwracać skale/skalowania
oznaczone jako do prezentacji (\code{TRUE} - tylko oznaczone jako do
prezentacji, \code{FALSE} - tylko oznaczone jako nie do prezentacji,
\code{NA} - wszystkie, bez względu na to, czy są do prezentacji, czy nie)}

\item{parametryzacja}{parametr określający format zwracanego wyniku. Domyślna
wartość to 'baza'. Inna możliwa wartość to 'mplus'.}
}
\value{
W przypadku użycia parametryzacji 'baza', funkcja zwraca listę taką,
że:
\itemize{
   \item{każdy element listy opisuje parametry innego skalowania;}
   \item{każdy element listy ma też przypisane jako atrybuty wartości kolumn:
         'skalowanie', 'opis', 'estymacja';}
   \item{zwracana lista ma przypisane jako atrybuty wartości wszystkich kolumn
         z tablicy 'skale';}
}
W przypadku użycia parametryzacji 'mplus' funkcja zwraca parametry
skalowania w formie ramki danych, która jest w postaci zwracanej przez
funkcję skaluj().
}
\description{
Funkcja pobiera parametry skalowań o podanej nazwie skali i numerze testu
oraz których opis skalowań spełnia podane wyrażenie regularne.
}

