% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pobierz_wartosci_wskaznikow.r
\name{pobierz_wartosci_wskaznikow}
\alias{pobierz_wartosci_wskaznikow}
\title{Pobieranie wartosci wskaznikow EWD.}
\usage{
pobierz_wartosci_wskaznikow(typSzkoly, lata, zapis = NULL, jst = NULL,
  idOke = FALSE, daneEWD = FALSE, daneAdresowe = TRUE,
  ladneNazwy = TRUE, fileEncoding = "windows-1250",
  zrodloDanychODBC = "EWD")
}
\arguments{
\item{typSzkoly}{ciąg znaków: 'gimn.', 'LO' lub 'T'}

\item{lata}{wektor liczb całkowitych - lata, dla których mają zostać pobrane wskaźniki
(wskaźniki identyfikowane są po ostatnim roku w ramach okresu, np. wskaźniki 2013-2011
uzyskamy podając jako argument \code{lata} wartość 2013)}

\item{zapis}{opcjonalnie nazwa pliku, do którego zostaną zapisane pobrane dane
(w formacie csv)}

\item{jst}{opcjonalnie wyrażenie regularne opisujące TERYTy JST, dla których mają
zostać zwrócone wskaźniki lub NULL, jeśli mają zostać zwrócone wartości wskaźników dla
całego kraju}

\item{idOke}{wartość logiczna (domyślnie FALSE) - czy dołączać kody OKE szkół?}

\item{daneEWD}{wartość logiczna (domyślnie FALSE) - czy dołączać informacje o wielkości
miejscowości i o tym, czy szkoła jest publiczna?}

\item{daneAdresowe}{wartość logiczna (domyślnie TRUE) - czy dołączać dane adresowe
szkół?}

\item{ladneNazwy}{wartość logiczna (domyślnie TRUE) - czy zmieniać nazwy kolumn na
pięknie opisowe?}

\item{fileEncoding}{ciąg znaków - strona kodowa, w której zostanie zapisany wynikowy
plik csv}

\item{zrodloDanychODBC}{opcjonalnie nazwa źródła danych ODBC, dającego dostęp do bazy
(domyślnie "EWD")}
}
\value{
data frame
}
\description{
Funkcja pobiera z bazy wartości wskaźników EWD. Domyślne wartości argumentów zostały
dobrane pod kątem wydawania wskazńików na zewnątrz.
}

