% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_wyniki_surowe.R
\name{pobierz_wyniki_surowe}
\alias{pobierz_wyniki_surowe}
\title{Pobieranie surowych wynikow egzaminu}
\usage{
pobierz_wyniki_surowe(
  rodzajEgzaminu = c("sprawdzian", "egzamin gimnazjalny", "matura",
    "egzamin ósmoklasisty"),
  lata = NULL,
  nadpisz = FALSE,
  daneKontekstowe = TRUE,
  src = NULL
)
}
\arguments{
\item{rodzajEgzaminu}{ciąg znaków - rodzaj egzaminu}

\item{lata}{NULL lub wektor liczb - lata, dla których mają zostać pobrane
wyniki; jeśli NULL pobrane zostaną wyniki z wszystkich dostępnych lat}

\item{nadpisz}{wartość logiczna - jeśli w miejscu, w którym mają być zapisane
wyniki są już pliki z wynikami, to czy je nadpisać? jeśli FALSE, pobrane
zostaną wyniki tylko z lat, dla których pliki jeszcze nie istnieją}

\item{daneKontekstowe}{wartość logiczna - czy pobrać również plik z danymi
kontekstowymi (tj. o uczniach i szkołach)?}

\item{src}{NULL połączenie z bazą danych IBE zwracane przez funkcję
\code{\link[ZPD]{polacz}}; pozwala posłużyć się połączeniem o wyższych niż
domyślne prawach dostępu, co ma znaczenie dla zakresu pobieranych danych
kontekstowych}
}
\value{
lista z nazwami zapisanych plików (niewidocznie)
}
\description{
Funkcja pobiera surowe wyniki egzaminu i zapisuje je na dysku w postaci
plików RData.
\itemize{
  \item{Funkcja sprawdzi, czy w aktywnym katalogu istnieje katalog
        \code{dane surowe}. Jeśli tak, zapisze tam pliki, jeśli nie, najpierw
        go utworzy, a potem zapisze tam pliki.}
  \item{Dane dotyczące wyników z poszczególnych lat zapisane zostaną
        w oddzielnych plikach, o nazwach postaci \code{nazwa egzaminu rok.RData}.}
  \item{Każdy plik w formacie RData będzie zawierał ramki danych z wynikami
        poszczególnych części egzaminu.
        \itemize{
          \item{Nazwy ramek danych odpowiadają prefiksom tych części egzaminu
                w bazie, w tablic\code{sl_czesci_egzaminow}.}
          \item{Ich wewnętrzna struktura jest taka sama, jak struktura
                wartości zwracanej przez funkcję pobierz_czesc_egzaminu()
                z pakietu ZPD.}
          \item{Mają przypisane klasy (oprócz \code{data.frame}):
                \code{czescEgzaminu} i \code{wynikiSurowe}.}
          \item{Mają przypisany atrybut \code{dataPobrania}, zawierający
                datę w formie ciągu znaków postaci rrrr-mm-dd.}
        }
       }
}
Dodatkowo, o ile parametr \code{daneKontekstowe} ma wartość \code{TRUE},
w tym samym katalogu zapisany zostanie plik \code{nazwa egzaminu - kontekstowe.RData},
zawierający ramkę danych o nazwie \code{prefiksKontekstowe} (gdzie
\code{prefiks} opisuje rodzaj egzaminu: \code{s} - sprawdzian, \code{g} - egz.
gimn., \code{m} - matura), będącą wynikiem wywołania funkcji
\code{\link{pobierz_dane_kontekstowe}}.
}
