% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/przygotuj_dane_do_ewd.R
\name{przygotuj_dane_do_ewd}
\alias{przygotuj_dane_do_ewd}
\title{Przygotowywanie danych do wyliczania modeli EWD}
\usage{
przygotuj_dane_do_ewd(katalogZDanymi, typSzkoly, lataDo, liczbaRocznikow = 1,
  wydluzenie = 1)
}
\arguments{
\item{katalogZDanymi}{ciąg znaków - ścieżka do katalogu, w którym znajdują
się pliki .RData z wynikami surowymi egzaminów i danymi kontekstowymi
o uczniach i szkołach}

\item{typSzkoly}{ciąg znaków: "gimn." | "LO" | "T"}

\item{lataDo}{wektor liczb całkowitych - lata egzaminu "na wyjściu", dla
których mają zostać przygotowane zbiory; w przypadku plików obejmujących'
kilka roczników absolwentów (tj. gdy \code{liczbaRocznikow > 1}) wskazuje
najpóźniejsze lata w ramach plików}

\item{liczbaRocznikow}{liczba całkowita - ile roczników absolwentów ma
obejmować pojedynczy plik (domyślnie 1 - zbiory "jednoroczne")}

\item{wydluzenie}{liczba całkowita - w przygotowanym zbiorze znajdą się
uczniowie o toku kształcenia wydłużonym maksymalnie o tyle lat}
}
\value{
wektor z nazwami zapisanych plików (niewidocznie)
}
\description{
Funkcja przygotowuje pliki z danymi do wyliczania modeli EWD,
wykorzystujących wyniki surowe lub znormalizowane ekwikwantylowo lub
wyskalowane - w zależności od tego, jakie znajdzie w katalogu podanym
argumentem \code{katalogZDanymi} - i zapisuje go w aktywnym katalogu.
}
\details{
Funkcja działa w oparciu o dane dostępne lokalnie, ściągnięte
wcześniej przy pomocy funkcji \code{\link{pobierz_wyniki_surowe}} lub
\code{\link{pobierz_wyniki_wyskalowane}}.

Normy #' ekwikwantylowe pobierane są z bazy, z użyciem funkcji
\code{\link[ZPD]{normalizuj}}.

Reguły wyboru, które wyniki zostaną zwrócone w sytuacji, gdy wykorzystywane
są wyniki wyskalowane i dla jakiejś skali dostępne są wyniki z kilku różńych
skalowań, opisane są w pomocy do funkcji
\code{\link{wczytaj_wyniki_wyskalowane}}.
}

