% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytywanie_danych_z_dysku.R
\name{wczytaj_dane_kontekstowe}
\alias{wczytaj_dane_kontekstowe}
\title{Funkcje pomocnicze przy przygotowywaniu danych do wyliczania modeli EWD}
\usage{
wczytaj_dane_kontekstowe(nazwaPliku, czyWyjscie = TRUE, lata = NULL)
}
\arguments{
\item{nazwaPliku}{ciąg znaków - nazwa pliku .RData}

\item{czyWyjscie}{wartość logiczna - czy zwrócić dane w zakresie odpowiednim
do łączenia z wynikami egzaminu "na wyjściu" (jeśli \code{FALSE} - zwrócone
zostaną w formacie odpowiednim do łączenia z wyikami egzaminu "na wejściu")}

\item{lata}{opcjonalnie wektor liczb całkowitych - lata egzaminu, do jakich
ma zostać zawężony zwracany zestaw danych}
}
\value{
data table
}
\description{
Funkcja wczytuje podany plik z danymi kontekstowymi i wybiera
z niego informacje w sposób odpowiedni do łączenia go potem z wynikami
egzaminu traktowanymi jako wyniki "na wejściu" lub wyniki "na wyjściu".
}

