% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytywanie_danych_z_dysku.R
\name{wczytaj_liczbe_przystepujacych}
\alias{wczytaj_liczbe_przystepujacych}
\title{Funkcje pomocnicze przy przygotowywaniu danych do wyliczania modeli EWD}
\usage{
wczytaj_liczbe_przystepujacych(nazwyPlikow)
}
\arguments{
\item{nazwyPlikow}{wektor ciągów znaków - nazwy plików z surowymi wynikami
egzaminu}
}
\value{
data table
}
\description{
Funkcja wczytuje dane z plików .RData z wynikami surowymi
egzaminu, i przygotowuje zestawienie opisujące, którzy zdający przystępowali
do których części egzaminu.
}

