% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/przygotuj_dane_surowe_do_ewd.R
\name{wczytaj_wyniki_egzaminu}
\alias{wczytaj_wyniki_egzaminu}
\title{Funkcje pomocnicze przy przygotowywaniu danych do wyliczania modeli EWD}
\usage{
wczytaj_wyniki_egzaminu(nazwyPlikow, daneKontekstowe = NULL)
}
\arguments{
\item{nazwyPlikow}{wektor ciągów znaków - nazwy plików z surowymi wynikami
egzaminu}

\item{daneKontekstowe}{opcjonalnie (acz typowo) data frame z danymi
kontekstowymi, typowo wynik wywołania funkcji
\code{\link{wczytaj_dane_kontekstowe}}}
}
\value{
data table
}
\description{
Funkcja wczytuje wyniki surowe egzaminu z podanych plików,
łączy je w jeden zbiór i łączy z podanymi danymi kontekstowymi.
}

