% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytywanie_danych_z_dysku.R
\name{wczytaj_wyniki_skalowania}
\alias{wczytaj_wyniki_skalowania}
\title{Funkcje pomocnicze przy przygotowywaniu danych do wyliczania modeli EWD}
\usage{
wczytaj_wyniki_skalowania(nazwyPlikow)
}
\arguments{
\item{nazwyPlikow}{wektor ciągów znaków - nazwy plików z surowymi wynikami
egzaminu}
}
\value{
data table
}
\description{
Funkcja wczytuje wyskalowane wyniki egzaminu z pliku .RData
z wynikami skalowania, zapisanego przez funkcje
\code{\link[EWDskale]{skaluj_spr}},
\code{\link[EWDskale]{skaluj_egz_gimn}} lub
\code{\link[EWDskale]{skaluj_matura}}.
}

