% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wczytywanie_danych_z_dysku.R
\name{wczytaj_wyniki_surowe}
\alias{wczytaj_wyniki_surowe}
\title{Funkcje pomocnicze przy przygotowywaniu danych do wyliczania modeli EWD}
\usage{
wczytaj_wyniki_surowe(nazwaPliku)
}
\arguments{
\item{nazwaPliku}{ciąg znaków - nazwa pliku z surowymi wynikami egzaminu}
}
\value{
data table
}
\description{
Funkcja wczytuje wyniki surowe egzaminu z pojedynczego pliku.
}

