% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytywanie_danych_z_dysku.R
\name{wczytaj_wyniki_wyskalowane}
\alias{wczytaj_wyniki_wyskalowane}
\title{Funkcje pomocnicze przy przygotowywaniu danych do wyliczania modeli EWD}
\usage{
wczytaj_wyniki_wyskalowane(nazwaPliku)
}
\arguments{
\item{nazwaPliku}{ciąg znaków - nazwa pliku z surowymi wynikami egzaminu}
}
\value{
data table
}
\description{
Funkcja wczytuje wyniki surowe egzaminu z pojedynczego pliku.
}
\details{
Jeśli plik dla jakiejś skali zawiera wyniki więcej niż jednego skalowania,
to zwrócone zostaną wyniki tylko jednego z nich, przy czym piwerwszeństwo
będą mieć skalowania oznaczone jako do prezentacji, a w dalszej kolejności
te, których data skalowania jest najpóżniejsza.
}

