% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weryfikuj_baze_szkol.R
\name{weryfikuj_baze_szkol}
\alias{weryfikuj_baze_szkol}
\title{Aktualizacja bazy danych szkol}
\usage{
weryfikuj_baze_szkol(bazaZakt, bazaZrzut = NULL)
}
\arguments{
\item{bazaZakt}{nazwa (ścieżka do) pliku w formacie csv
zawierającego zaktualizowaną bazę szkół lub data frame z taką wczytaną bazą}

\item{bazaZrzut}{opcjonalnie nazwa (ścieżka do) pliku w formacie csv
zawierającego zrzut bazy szkół, na który nanoszona była aktualizacja lub data
frame z taką wczytaną bazą}
}
\value{
lista data frame'ów
}
\description{
Funkcja weryfikuje poprawność bazy szkół, przygotowanej w pliku csv, po
przeprowadzeniu jej ręcznej aktualizacji.
}
\details{
Jeśli parametr \code{bazaZrzut} nie zostanie podany, funkcja jedynie
zweryfikuje poprawność podanej bazy. Jeśli zostanie on podany, o ile nie
zostaną wykryte błędy polegające na duplikacji identyfikatorów OKE, funkcja
zbada również zakres zmian w stosunku do zrzutu bazy.
}

