% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wielokrotnie_przystepujacy.r
\name{wielokrotnie_przystepujacy}
\alias{wielokrotnie_przystepujacy}
\title{Uczniowie wielokrotnie przystępujący do egzaminu.}
\usage{
wielokrotnie_przystepujacy(rodzajEgzaminu, zrodloDanychODBC = "EWD")
}
\arguments{
\item{rodzajEgzaminu}{rodzaj egzaminu ("sprawdzian" / "egzamin gimnazjalny" / "matura")}

\item{zrodloDanychODBC}{opcjonalnie nazwa źródła danych ODBC, dającego dostęp do bazy (domyślnie "EWD")}
}
\value{
Data frame, który zawiera \code{id_obserwacji} oraz pierwszy i ostatni rok, w którym dany uczeń przystąpił do egzaminu.
W zwracanym obiekcie występują tylko ci uczniowie, którzy podchodzili do egzaminu wielokrotnie (tj. pierwszy rok przystąpienia nie jest dla nich równy ostatniemu rokowi przystąpienia).
}
\description{
Funkcja zwraca informacje o uczniach wielokrotnie przystępujących do egzaminu.
}
\details{
Wszystkie \code{id_testu} podane w argumencie \code{idTestow} muszą być powiązane z tym samym egzaminem.
Wiek zwracany jest w miesiącach. Jeśli testy były pisane w różnych dniach, zwrócony zostanie wiek wyliczony w odniesieniu do daty testu, który został przeprowadzony najwcześniej.
}

