% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_parametry_skalowania.R
\name{zmien_na_mplus}
\alias{zmien_na_mplus}
\title{Zmiana tablicy do formatu funkcji skaluj()}
\usage{
zmien_na_mplus(x)
}
\arguments{
\item{x}{data frame (data table) w formacie takim, jaki zwraca funkcja
\code{\link[ZPD]{pobierz_parametry}} z pakietu \code{ZPD}}
}
\value{
Funkcja zwraca ramkę danych, która jest zgodna z postacią ramek zwracanych
przez funkcję \code{\link[EWDskalowanie]{skaluj}}.
}
\description{
Funkcja przekształca parametryzację z bazy (klasyczne, jednowymiarowe IRT)
na parametryzację stosowaną przez Mplusa.
}
