% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki_ewd.R
\name{agreguj_wskazniki_ewd}
\alias{agreguj_wskazniki_ewd}
\title{Agregowanie wskaznikow EWD}
\usage{
agreguj_wskazniki_ewd(dane, poziom = NULL, grupujPoLatach = TRUE,
  funkcjaGrupujaca = NULL, tylkoWyswietlane = NA, paramGrupujaca = list())
}
\arguments{
\item{dane}{ramka danych zwracana przez funkcję \link{pobierz_wartosci_wskaznikow_ewd}.}

\item{poziom}{ciąg znaków definiujący poziom agregacji: "gmina", "powiat" lub
"województwo", albo \code{NULL}, jeśli podany został parametr 'funkcjaGrupujaca'}

\item{grupujPoLatach}{wartość logiczna (\code{TRUE} lub \code{FALSE})
wskazująca, czy przy wyróżnianiu grup oprócz TERYTu ma być uwzględniona
również zmienna opisująca okres, dla którego wyliczony został wskaźnik}

\item{funkcjaGrupujaca}{NULL lub fukcja przyjmująca jako pierwszy parametr
TERYT szkoły (w postaci wektora liczb całkowitych) i zwracająca jednoelementowy
data frame (lub listę), który opisuje przydział szkół do grup}

\item{tylkoWyswietlane}{wartość logiczna (\code{TRUE/FALSE/NA}) opisująca,
czy przy wyliczaniu agregatów mają zostać uwzględnione tylko te szkoły,
których elipsy są pokazywane na stronie - p. sekcja Details}

\item{paramGrupujaca}{opcjonalnie lista dodatkowych parametrów, które zostaną
przekazane funkcji grupującej}
}
\value{
data frame
}
\description{
Funkcja agreguje wskaźniki EWD.
}
\details{
Zachowanie funkcji w zależności od wartości parametru \code{tylkoWyswietlane}:
\itemize{
  \item{Jeśli \code{tylkoWyswietlane=TRUE} przy wyliczaniu agregatów
        uwzględniane zostaną tylko te szkoły, których elipsy są wyświetlane
        na stronie.}
  \item{Jeśli \code{tylkoWyswietlane=FALSE} przy wyliczaniu agregatów
        uwzględniane zostaną wszystkie szkoły}
  \item{Domyślnie, tj. gdy \code{tylkoWyswietlane=NA}, przy wyliczaniu
        agregatów uwzględniane zostaną wszystkie szkoły, (jak wtedy, gdy
        \code{tylkoWyswietlane=TRUE}) z tym, że jeśli w jakiejś grupie nie ma
        żadnej szkoły, której elipsa byłaby prezentowana, to wartość
        wskaźnika zostanie w niej zmieniona na brak danych.}
}
Oznacza to, że dla poprawnego działania domyślnego wywołwania funkcji
\code{agreguj_wskazniki_ewd} potrzebne jest pobranie wcześniej danych
o wartościach wskaźników wywołaniem funkcji
\code{\link{pobierz_wartosci_wskaznikow_ewd}} z inną niż domyślna wartością
parametru \code{tylkoWyswietlane} - p. przykład użycia.
}
\examples{
\dontrun{
dane = pobierz_wartosci_wskaznikow_ewd("T", 2013:2014, tylkoWyswietlane = FALSE) \%>\%
  agreguj_wskazniki_ewd("powiat")
}
}

