% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobierz_dane_kontekstowe.R
\name{pobierz_dane_kontekstowe}
\alias{pobierz_dane_kontekstowe}
\title{Pobieranie danych kontekstowych o uczniach.}
\usage{
pobierz_dane_kontekstowe(src, rodzajEgzaminu)
}
\arguments{
\item{src}{połączenie z bazą danych IBE zwracane przez ZPD::polacz()}

\item{rodzajEgzaminu}{rodzaj egzaminu}
}
\value{
data.frame
}
\description{
Pobiera z bazy danych zmaterializowany zbiór danych zawierający wszelkie
infromacje kontekstowe potrzebne przy wyliczaniu EWD.

Jeden wiersz zwracanej ramki danych opisuje pierwsze/ostatnie (patrz niżej)
przystąpienie ucznia do wskazanego rodzaju egzaminu, a informacje dotyczące
poszczególnych części egzaminu (na chwilę obecną tylko informacja o byciu
laureatem z danego przedmiotu) znajdują się w kolumnach prefiksowanych
zgodnie z tablicą sl_czesci_egzaminow w bazie.

W zależności od wartości parametru \code{wyjscie} pobierane są informacje o
uczniach podchodzących po raz pierwszy (\code{wyjscie == TRUE}) lub po raz
ostatni (\code{wyjscie == FALSE}) do wskazanego rodzaju egzaminu.

Jeśli informacje o szkole, klasie, numerze w dzienniku lub dysleksji nie są
dla danego ucznia zgodne w różnych częściach egzaminu, wtedy dana informacja
(szkoła, klasa, dysleksja) zastępowana jest brakiem danych.

W zależności od poziomu uprawnień połączenia z bazą danych funkcja
automatycznie pobiera lub nie informacje o kodzie klasy i numerze ucznia w
dzienniku.

Dane pobierane są w postaci zmaterializowanej z powodów wydajnościowych -
zapytanie realizujące całą niezbędną funkcjonalnośc po stronie bazy danych
wykonywałoby się wieki.
}

