% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_parametry_skalowania.R
\name{pobierz_parametry_skalowania}
\alias{pobierz_parametry_skalowania}
\title{Pobieranie parametrow skalowan}
\usage{
pobierz_parametry_skalowania(
  skala,
  skalowanie = NULL,
  doPrezentacji = NA,
  parametryzacja = "baza",
  src = NULL
)
}
\arguments{
\item{skala}{id_skali (liczba naturalna) lub ciąg znaków z wyrażeniem
regularnym, do którego ma pasować opis skali}

\item{skalowanie}{opcjonalnie numer skalowania (liczba naturalna) lub ciąg
znaków z wyrażeniem regularnym, do którego ma pasować opis skalowania}

\item{doPrezentacji}{wartość logiczna - czy zwracać skale/skalowania
oznaczone jako do prezentacji (\code{TRUE} - tylko oznaczone jako do
prezentacji, \code{FALSE} - tylko oznaczone jako nie do prezentacji,
\code{NA} - wszystkie, bez względu na to, czy są do prezentacji, czy nie)}

\item{parametryzacja}{parametr określający format zwracanego wyniku. Domyślna
wartość to 'baza'. Inna możliwa wartość to 'mplus'.}

\item{src}{NULL połączenie z bazą danych IBE zwracane przez funkcję
\code{\link[ZPD]{polacz}}. Jeśli nie podane, podjęta zostanie próba
automatycznego nawiązania połączenia.}
}
\value{
data frame (data table) o kolumnach:
\itemize{
   \item{id_skali;}
   \item{rodzaj_egzaminu;}
   \item{opis_skali;}
   \item{skalowaie;}
   \item{parametry.}
}
Ostatnie kolumna zawiera data frame'y z wartościami parametrów skalowań.
W zależności od wartości argumentu \code{parametryzacja} albo w formacie
zwracanym przez funkcję \code{\link[ZPD]{pobierz_parametry}}, albo w formacie
wykorzystywanym przez funkcję \code{\link[EWDskalowanie]{skaluj}}.
}
\description{
Funkcja pobiera parametry skalowań o podanej nazwie skali i numerze testu
oraz których opis skalowań spełnia podane wyrażenie regularne.
}
