% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/agreguj_wskazniki_ewd.R
\name{przygotuj_funkcje_grupujaca_teryt}
\alias{przygotuj_funkcje_grupujaca_teryt}
\title{Przygotowanie funkcji grupujacej w JST na podstawie TERYTu}
\usage{
przygotuj_funkcje_grupujaca_teryt(poziom)
}
\arguments{
\item{poziom}{ciąg znaków definiujący poziom agregacji: "gmina", "powiat" lub
"województwo"}
}
\value{
funkcja
}
\description{
Funkcja przygotowuje funkcję grupującą. Więcej szczegółów
w \link{agreguj_wskazniki_ewd}
}

