% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wczytywanie_danych_z_dysku.R
\name{wczytaj_wyniki_egzaminu}
\alias{wczytaj_wyniki_egzaminu}
\title{Funkcje pomocnicze przy przygotowywaniu danych do wyliczania modeli EWD}
\usage{
wczytaj_wyniki_egzaminu(nazwyPlikow, daneKontekstowe = NULL)
}
\arguments{
\item{nazwyPlikow}{wektor ciągów znaków - nazwy plików z surowymi wynikami
egzaminu}

\item{daneKontekstowe}{opcjonalnie (acz typowo) data frame z danymi
kontekstowymi, typowo wynik wywołania funkcji
\code{\link{wczytaj_dane_kontekstowe}}}
}
\value{
data table
}
\description{
Funkcja wczytuje wyniki egzaminu (surowe lub wyskalowane)
z podanych plików, łączy je w jeden zbiór i ewentualnie łączy z podanymi
danymi kontekstowymi.

Reguły wyboru, które wyniki zostaną zwrócone w sytuacji, gdy wykorzystywane
są wyniki wyskalowane i dla jakiejś skali dostępne są wyniki z kilku różńych
skalowań, opisane są w pomocy do funkcji
\code{\link{wczytaj_wyniki_wyskalowane}}.
}

