% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobierz_parametry_skalowania.R
\name{zmien_na_mplus}
\alias{zmien_na_mplus}
\title{Zmiana tablicy do formatu funkcji skaluj()}
\usage{
zmien_na_mplus(tablicaDanych)
}
\arguments{
\item{tablicaDanych}{tablica w formacie zwracanym przez funkcję
\code{\link{pobierz_parametry_skalowania}}.}
}
\value{
Funkcja zwraca ramkę danych, która jest zgodna z postacią ramek zwracanych przez
funkcję \code{\link[EWDskalowanie]{skaluj}}.
}
\description{
Funkcja przekształca tablicę zwracaną w liście przez funkcję
\code{\link{pobierz_parametry_skalowania}} z parametrem 'baza' do postaci mplusa
}

